+++ libc-bottom-half/cloudlibc/src/libc/sys/ioctl/ioctl.c
@@ -4,7 +4,7 @@
 
 #include <sys/ioctl.h>
 
-#include <wasi/api.h>
+#include <wasix/api.h>
 #include <errno.h>
 #include <stdarg.h>
 
@@ -23,7 +23,7 @@ int ioctl(int fildes, int request, ...) {
           },
       };
       __wasi_event_t events[__arraycount(subscriptions)];
-      size_t nevents;
+      uint32_t nevents;
       __wasi_errno_t error = __wasi_poll_oneoff(
           subscriptions, events, __arraycount(subscriptions), &nevents);
       if (error != 0) {
