+++ libc-bottom-half/cloudlibc/src/libc/sys/socket/send.c
@@ -7,7 +7,7 @@
 #include <sys/socket.h>
 
 #include <assert.h>
-#include <wasi/api.h>
+#include <wasix/api.h>
 #include <errno.h>
 
 ssize_t send(int socket, const void *buffer, size_t length, int flags) {
@@ -24,7 +24,7 @@ ssize_t send(int socket, const void *buffer, size_t length, int flags) {
   __wasi_siflags_t si_flags = 0;
 
   // Perform system call.
-  size_t so_datalen;
+  uint64_t so_datalen;
   __wasi_errno_t error = __wasi_sock_send(socket, si_data, si_data_len, si_flags, &so_datalen);
   if (error != 0) {
     errno = errno_fixup_socket(socket, error);
