+++ libc-bottom-half/cloudlibc/src/libc/time/clock_nanosleep.c
@@ -6,7 +6,7 @@
 #include <common/time.h>
 
 #include <assert.h>
-#include <wasi/api.h>
+#include <wasix/api.h>
 #include <errno.h>
 #include <time.h>
 
@@ -28,7 +28,7 @@ int clock_nanosleep(clockid_t clock_id, int flags, const struct timespec *rqtp,
     return EINVAL;
 
   // Block until polling event is triggered.
-  size_t nevents;
+  uint32_t nevents;
   __wasi_event_t ev;
   __wasi_errno_t error = __wasi_poll_oneoff(&sub, &ev, 1, &nevents);
   return error == 0 && ev.error == 0 ? 0 : ENOTSUP;
