+++ libc-bottom-half/cloudlibc/src/libc/unistd/read.c
@@ -2,13 +2,13 @@
 //
 // SPDX-License-Identifier: BSD-2-Clause
 
-#include <wasi/api.h>
+#include <wasix/api.h>
 #include <errno.h>
 #include <unistd.h>
 
 ssize_t read(int fildes, void *buf, size_t nbyte) {
   __wasi_iovec_t iov = {.buf = buf, .buf_len = nbyte};
-  size_t bytes_read;
+  uint64_t bytes_read;
   __wasi_errno_t error = __wasi_fd_read(fildes, &iov, 1, &bytes_read);
   if (error != 0) {
     errno = error == ENOTCAPABLE ? EBADF : error;
